Fixed minor errors in reducers
