+++ app/actions/lab.js
@@ -3,9 +3,9 @@ import db from '../db';
 const LAB_CREATE = 'LAB_CREATE';
 const LAB_CREATED = 'LAB_CREATED';
 
-export const labCreate = fields => async dispatch => {
-  // TODO: Get real password
-  const { Lab } = db.initialize(7, '123');
+export const labCreate = fields => async (dispatch, getState) => {
+  const { user } = getState();
+  const { Lab } = db.initializeForUser(user);
   dispatch({ type: LAB_CREATE, fields });
 
   await Lab.create(fields);
