+++ app/actions/labs.js
@@ -1,40 +1,19 @@
 import db from '../db';
+import { fetchPaginated } from '../utils/fetch';
 
 const FETCH_LABS = 'FETCH_LABS';
 const FETCH_LABS_FAILED = 'FETCH_LABS_FAILED';
 
 // TODO: Abstract this to a helper function
-const fetchPaginated = async (url, auth) => {
-  const f = page =>
-    fetch(`${url}?page=${page}`, {
-      headers: auth
-    }).then(res => res.json());
-
-  // This gets the first page, so we can iterate over each page
-  const { items: firstItems, total_pages: totalPages } = await f(1);
-
-  if (totalPages <= 1) return Promise.resolve(firstItems);
-  // After that, we iterate over the following pages and save the result into an array
-  // This generates 1 request per page
-  const otherPages = (await Promise.all(
-    [...new Array(totalPages - 1)].map(async (_, i) => {
-      const { items } = await f(i + 2);
-      return items;
-    })
-  )).reduce((arr, acc) => [...acc, ...arr], []);
-  const items = [...firstItems, ...otherPages];
-  return Promise.resolve(items);
-};
-
 const labMapper = props => ({
   ...props,
   remoteId: props.id
 });
 
-export const fetchLabs = () => async (dispatch, store) => {
-  dispatch({ type: FETCH_LABS });
-  const { Lab } = db.initialize('asd', '123');
-  fetchPaginated('http://localhost:3000/api/v1/labs', store().user.auth)
+export const fetchLabs = () => async (dispatch, getState) => {
+  const { user } = getState();
+  const { Lab } = db.initializeForUser(user);
+  fetchPaginated('http://localhost:3000/api/v1/labs', user.auth)
     .then(res =>
       res.map(async item => {
         const mapper = labMapper(item);
